requirements.txt
aiosqlite==0.21.0
alembic==1.16.2
anyio==4.9.0
APScheduler==3.11.0
certifi==2025.6.15
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
pytest==8.4.1
pytest-asyncio==1.0.0
python-dotenv==1.1.1
python-telegram-bot==22.2
sniffio==1.3.1
SQLAlchemy==2.0.41
typing_extensions==4.14.0
tzlocal==5.3.1
